Reworked Viewport handling.

+++ libtitania-x3d/Titania/X3D/Browser/Navigation/X3DExamineViewer.h
@@ -172,7 +172,7 @@ private:
 	time_type        pressTime;
 	time_type        motionTime;
 	guint            button;
-	int              numClicks;
+	size_t           numClicks;
 	sigc::connection spin_id;
 
 };
